Configured Main class to populate database on startup
Modified the Main class to include a CommandLineRunner bean, which executes on application startup. This bean interacts with the CustomerRepository to save initial customer data directly to the database. Customer objects for Alex and Jamila are created, and a List of customers is saved using the saveAll() method of the repository.

This configuration ensures that customer data is automatically added to the database when the application starts, streamlining the initialization of data.